- regular expression to filter interaction box in item data corrected to prevent an item section to be removed - confirm box added on interaction deleting
git-svn-id: http://vcs.taotesting.com/svn/oat/projects/taoItems/trunk@3917 62dfeb86-a93f-4914-9d13-15c0a160985d

+++ views/js/qtiAuthoring/qtiEditClass.js
@@ -341,7 +341,9 @@ qtiEdit.prototype.bindInteractionLinkListener = function(editorDoc){
 		$deleteButton = $('<span class="qti_interaction_box_delete"></span>').appendTo($interactionContainer);
 		$deleteButton.hide();
 		$deleteButton.bind('click', {'interactionSerial': interactionSerial}, function(e){
-			instance.deleteInteractions([e.data.interactionSerial]);
+			if(confirm(__('Please confirm interaction deletion'))){
+				instance.deleteInteractions([e.data.interactionSerial]);
+			}
 		});
 		qtiEdit.makeNoEditable($interaction.parent());
